requirements.txt
# Core Framework
fastapi==0.119.0
uvicorn==0.37.0

# Database
sqlalchemy==2.0.44
psycopg2-binary==2.9.11

# Data Validation
pydantic==2.12.2
pydantic-settings==2.6.1

# Environment Variables
python-dotenv==1.1.1

# Testing
pytest==8.3.3
httpx==0.27.2

# Dependências instaladas automaticamente
# annotated-types==0.7.0
# anyio==4.11.0
# click==8.3.0
# colorama==0.4.6
# greenlet==3.2.4
# h11==0.16.0
# idna==3.11
# pydantic_core==2.41.4
# sniffio==1.3.1
# starlette==0.48.0
# typing-inspection==0.4.2
# typing_extensions==4.15.0